fixing light source

+++ src/main/java/at/digitalphoenix/firstmod/blocks/fire/FireLamp.java
@@ -5,7 +5,7 @@ import net.minecraft.block.material.Material;
 
 public class FireLamp extends LightSource{
 
-	public FireLamp(String name, Material material, float lightLevel) {
+	public FireLamp(String name, Material material, int lightLevel) {
 		super(name, material, lightLevel);
 		setHardness(2.0f);
 		setHarvestLevel("axe", 1);
